pyproject.toml
[tool.poetry]
name = "scroogebot"
version = "0.1.0"
description = ""
authors = ["blcklptn"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.4"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.1"
pydantic = "^1.10.5"
aiogram = "3.0b3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
